+++ src/main/java/chess/domain/board/Board.java
@@ -7,10 +7,10 @@ import chess.domain.chess.Color;
 import chess.domain.chess.Status;
 import chess.domain.piece.Blank;
 import chess.domain.piece.Piece;
-import chess.domain.piece.PieceDto;
 import chess.domain.piece.PieceFactory;
 import chess.domain.position.MovePosition;
 import chess.domain.position.Position;
+import chess.web.dto.PieceDto;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
