+++ src/main/java/chess/web/controller/MainController.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.web.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
