+++ src/main/java/chess/web/dao/ChessDao.java
@@ -1,21 +1,21 @@
-package chess.domain.chess;
+package chess.web.dao;
 
-import chess.domain.piece.PieceDto;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
+import chess.domain.chess.Chess;
+import chess.web.dto.PieceDto;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public class ChessDao {
 
-    JdbcTemplate jdbcTemplate;
+    private final JdbcTemplate jdbcTemplate;
 
     public ChessDao(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
@@ -23,7 +23,7 @@ public class ChessDao {
 
     public Chess findChessById(Long chessId) {
         String sql = "SELECT c.status, c.turn, p.position, p.color, p.name "
-                + "FROM chess c LEFT JOIN piece p ON c.chess_id = p.chess_id "
+                + "FROM chess AS c LEFT JOIN piece AS p ON c.chess_id = p.chess_id "
                 + "WHERE c.chess_id = (?)";
 
         return jdbcTemplate.queryForObject(sql, chessMapper(), chessId);
@@ -60,21 +60,11 @@ public class ChessDao {
     }
 
     public long insert() {
-        String sql = "INSERT INTO chess (status, turn) VALUES ('RUNNING', 'WHITE')";
-        try (Connection connection = jdbcTemplate.getDataSource().getConnection();
-                PreparedStatement preparedStatement = connection
-                        .prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-                ResultSet rs = preparedStatement.getGeneratedKeys()) {
-
-            preparedStatement.executeUpdate();
-
-            if (rs.next()) {
-                return rs.getLong(1);
-            }
-        } catch (SQLException throwable) {
-            throwable.printStackTrace();
-        }
-        throw new IllegalStateException("체스 게임 생성에 실패했습니다.");
+        final KeyHolder keyHolder = new GeneratedKeyHolder();
+        final String sql = "INSERT INTO chess (status, turn) VALUES ('RUNNING', 'WHITE')";
+        jdbcTemplate.update(con -> con.prepareStatement(
+                sql, new String[]{"chess_id"}), keyHolder);
+        return keyHolder.getKey().longValue();
     }
