+++ src/test/java/chess/domain/chess/ColorTest.java
@@ -1,8 +1,7 @@
-package chess.domain;
+package chess.domain.chess;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.chess.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
